requirements.txt
fastapi
uvicorn
openai
ibmcloudant
numpy
python-dotenv
faiss-cpu
